+++ config/reporter.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+
+    'default_driver' => 'pdf',
+
+    'drivers' => [
+        'pdf' => null,
+        'xlsx' => null,
+        'csv' => null,
+    ],
+
+];