+++ src/Contracts/Driver.php
@@ -6,7 +6,5 @@ use Lageg\Reporter\Report;
 
 interface Driver
 {
-    public function config(array $config = []): self;
-    public function generate(Exporter $exporter): Report;
-    public function chunkrize(Chunkrizable $chunkrizable): Report;
+    public function generate(Exportable $exportable, array $config): Report;
 }