+++ CHANGELOG.md
@@ -5,13 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2026-03-11
+
+### Added
+
+- Drivers alias config.
+- Contracts\Exportable interface.
+- Exporter base class.
+- Reporter Drive Manager.
+
+### Fixed
+
+- Report output path uses original filename.
+
+### Removed
+
+- Contracts\Exporter interface.
+- Contracts\Chunkrizable interface.
+- Traits\HasComponents trait.
+- Jobs\GenerateReportJob job.
+
 ## [1.1.0] - 2026-03-10
 
 ### Added
 
-- Traits\HasComponents Trait
-- CHANGELOG.md
+- Traits\HasComponents Trait.
+- CHANGELOG.md.
 
 ### Removed
 
-- Trats\ProvidesComponents Trait
+- Trats\ProvidesComponents Trait.